requirements.txt
requests
python-dotenv
langchain-groq
langchain-core
streamlit
speechrecognition
pydub
langdetect
deep-translator
gtts